requirements.txt
typer>=0.12.3
PyYAML>=6.0.1
python-dotenv>=1.0.1
openai>=1.30.0
chromadb>=0.5.0
